fix: #27 ReportFeature내부에 있는 Child 파일 분리
